split PD and gemma4 pp tests onto 2-gpu by GPU footprint
