+++ test/registered/disaggregation/test_disaggregation_kimi_linear.py
@@ -14,7 +14,7 @@ from sglang.test.test_utils import (
     popen_launch_server,
 )
 
-register_cuda_ci(est_time=1380, stage="base-c", runner_config="4-gpu-h100")
+register_cuda_ci(est_time=480, stage="base-c", runner_config="4-gpu-h100")
 
 KIMI_LINEAR_MODEL = "yujiepan/kimi-linear-tiny-random"
 SERVER_ENV = {"SGLANG_BATCH_INVARIANT_OPS_ENABLE_MM_DEEPGEMM": "0"}
@@ -34,30 +34,8 @@ DETERMINISTIC_ARGS = [
     "disabled",
 ]
 
-# --attention-backend and --enable-deterministic-inference are deliberately
-# absent: bytes are backend-independent and both paths run identical shapes.
-UNIFIED_MEMORY_ARGS = [
-    "--skip-tokenizer-init",
-    "--random-seed",
-    "1",
-    "--enable-unified-memory",
-    # Page-major layout requires triton for both.
-    "--linear-attn-backend",
-    "triton",
-    "--mamba-backend",
-    "triton",
-    "--max-mamba-cache-size",
-    "32",
-    "--max-total-tokens",
-    "4096",
-    "--cuda-graph-backend-decode",
-    "disabled",
-    "--cuda-graph-backend-prefill",
-    "disabled",
-]
 
-
-class _KimiLinearPDParityBase(PDDisaggregationServerBase):
+class KimiLinearPDParityBase(PDDisaggregationServerBase):
     reference_parallel_args = []
     baseline_args = []
     extra_prefill_env = SERVER_ENV
@@ -123,7 +101,7 @@ class _KimiLinearPDParityBase(PDDisaggregationServerBase):
         assert_process_healthy(self, "decode", self.process_decode, self.decode_url)
 
 
-class TestKimiLinearHeterogeneousTPDisaggregation(_KimiLinearPDParityBase):
+class TestKimiLinearHeterogeneousTPDisaggregation(KimiLinearPDParityBase):
     prefill_tp_size = 2
     decode_tp_size = 1
     decode_base_gpu_id = 2
@@ -139,32 +117,6 @@ class TestKimiLinearPipelineDisaggregation(TestKimiLinearHeterogeneousTPDisaggre
     decode_base_gpu_id = 2
     reference_parallel_args = ["--tp-size", "1", "--pp-size", "2"]
     extra_prefill_args = DETERMINISTIC_ARGS + ["--pp-size", "2"]
-
-
-class TestUnifiedMemoryDisaggregation(_KimiLinearPDParityBase):
-    """1 prefill + 1 decode, both with --enable-unified-memory, vs a non-PD
-    unified-memory reference server."""
-
-    prefill_tp_size = 1
-    decode_tp_size = 1
-    decode_base_gpu_id = 1
-    baseline_args = UNIFIED_MEMORY_ARGS
-    extra_prefill_args = UNIFIED_MEMORY_ARGS
-    extra_decode_args = UNIFIED_MEMORY_ARGS
-
-
-class TestUnifiedMemoryDisaggregationChunkedPrefill(TestUnifiedMemoryDisaggregation):
-    """Multi-chunk prefill (257-token prompt, 64-token chunks): each chunk's KV
-    pages are translated to physical ids and shipped while later chunks still
-    run, exercising the chunked send path and the prefill-side move gate
-    (`chunked_req.start_send_idx > 0`). The reference server uses the same
-    chunk size so any parity break isolates to the PD transfer.
-    """
-
-    _chunked_args = UNIFIED_MEMORY_ARGS + ["--chunked-prefill-size", "64"]
-    baseline_args = _chunked_args
-    extra_prefill_args = _chunked_args
-    extra_decode_args = _chunked_args
 
 
 if __name__ == "__main__":
